requirements.txt
time
dash
plotly.graph_objs
wmi
argparse
subprocess
random
psutil
threading
os
